Refactor(plugins): Move filter snmp_hash to PyAVD
