+++ python-avd/pyavd/_eos_designs/structured_config/base/snmp_server.py
@@ -8,8 +8,8 @@ from hashlib import sha1
 from typing import TYPE_CHECKING
 
 from ....j2filters.natural_sort import natural_sort
+from ....j2filters.snmp_hash import snmp_hash
 from ....vendor.errors import AristaAvdError, AristaAvdMissingVariableError
-from ....vendor.j2.filter.snmp_hash import hash_passphrase
 from ....vendor.strip_empties import strip_null_from_data
 from ....vendor.utils import get, replace_or_append_item
 from .utils import UtilsMixin
@@ -33,7 +33,7 @@ class SnmpServerMixin(UtilsMixin):
         unique local_engine_id value based on hostname and mgmt_ip facts.
 
         If user.version is set to 'v3', compute_local_engineid and compute_v3_user_localized_key are set to 'True'
-        we will use hash_passphrase filter to create an instance of hashlib._hashlib.HASH corresponding to the auth_type
+        we will use snmp_hash filter to create an instance of hashlib HASH corresponding to the auth_type
         value based on various snmp_settings.users information.
         """
         source_interfaces_inputs = self._source_interfaces.get("snmp")
@@ -138,7 +138,7 @@ class SnmpServerMixin(UtilsMixin):
                     user_dict["auth"] = auth
                     if compute_v3_user_localized_key:
                         hash_filter = {"passphrase": auth_passphrase, "auth": auth, "engine_id": engine_ids["local"]}
-                        user_dict["auth_passphrase"] = hash_passphrase(hash_filter)
+                        user_dict["auth_passphrase"] = snmp_hash(hash_filter)
                     else:
                         user_dict["auth_passphrase"] = auth_passphrase
 
@@ -148,7 +148,7 @@ class SnmpServerMixin(UtilsMixin):
                         user_dict["priv"] = priv
                         if compute_v3_user_localized_key:
                             hash_filter.update({"passphrase": priv_passphrase, "priv": priv})
-                            user_dict["priv_passphrase"] = hash_passphrase(hash_filter)
+                            user_dict["priv_passphrase"] = snmp_hash(hash_filter)
                         else:
                             user_dict["priv_passphrase"] = priv_passphrase
