+++ python-avd/pyavd/_utils/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2024 Arista Networks, Inc.
+# Use of this source code is governed by the Apache License 2.0
+# that can be found in the LICENSE file.
